move dns-discovery
