+++ packages/discovery/package.json
@@ -51,10 +51,28 @@
     "node": ">=18"
   },
   "dependencies": {
+    "@waku/enr": "0.0.20",
+    "@waku/utils": "0.0.14",
+    "debug": "^4.3.4",
+    "dns-query": "^0.11.2",
+    "hi-base32": "^0.5.1",
+    "uint8arrays": "^5.0.1"
+  },
+  "devDependencies": {
+    "@libp2p/peer-id": "^4.0.4",
+    "@libp2p/peer-id-factory": "^4.0.5",
+    "@multiformats/multiaddr": "^12.0.0",
     "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-json": "^6.0.0",
     "@rollup/plugin-node-resolve": "^15.2.3",
-    "@waku/build-utils": "^1.0.0"
+    "@types/chai": "^4.3.11",
+    "@waku/build-utils": "*",
+    "@waku/interfaces": "0.0.21",
+    "chai": "^4.3.10",
+    "cspell": "^8.3.2",
+    "mocha": "^10.3.0",
+    "npm-run-all": "^4.1.5",
+    "rollup": "^4.12.0"
   },
   "files": [
     "dist",
