+++ packages/discovery/src/dns/constants.ts
