+++ packages/discovery/src/dns/dns.ts
