+++ packages/discovery/src/dns/index.spec.ts
