+++ packages/discovery/src/dns/testdata.json
