--- packages/dns-discovery/.eslintrc.cjs
@@ -1,6 +1,0 @@
-module.exports = {
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: "./tsconfig.dev.json"
-  }
-};
