--- packages/dns-discovery/karma.conf.cjs
@@ -1,3 +1,0 @@
-const config = require("../../karma.conf.cjs");
-
-module.exports = config;
