--- packages/dns-discovery/package.json
@@ -1,87 +1,0 @@
-{
-  "name": "@waku/dns-discovery",
-  "version": "0.0.20",
-  "description": "DNS Peer Discovery (EIP-1459)",
-  "types": "./dist/index.d.ts",
-  "module": "./dist/index.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "type": "module",
-  "author": "Waku Team",
-  "homepage": "https://github.com/waku-org/js-waku/tree/master/packages/dns-discovery#readme",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/waku-org/js-waku.git"
-  },
-  "bugs": {
-    "url": "https://github.com/waku-org/js-waku/issues"
-  },
-  "license": "MIT OR Apache-2.0",
-  "keywords": [
-    "waku",
-    "decentralized",
-    "secure",
-    "communication",
-    "web3",
-    "ethereum",
-    "dapps",
-    "privacy"
-  ],
-  "scripts": {
-    "build": "run-s build:**",
-    "build:esm": "tsc",
-    "build:bundle": "rollup --config rollup.config.js",
-    "fix": "run-s fix:*",
-    "fix:lint": "eslint src *.js --fix",
-    "check": "run-s check:*",
-    "check:lint": "eslint src --ext .ts",
-    "check:spelling": "cspell \"{README.md,src/**/*.ts}\"",
-    "check:tsc": "tsc -p tsconfig.dev.json",
-    "prepublish": "npm run build",
-    "reset-hard": "git clean -dfx -e .idea && git reset --hard && npm i && npm run build",
-    "test": "NODE_ENV=test run-s test:*",
-    "test:node": "NODE_ENV=test TS_NODE_PROJECT=./tsconfig.dev.json mocha",
-    "test:browser": "NODE_ENV=test karma start karma.conf.cjs"
-  },
-  "engines": {
-    "node": ">=18"
-  },
-  "dependencies": {
-    "@waku/enr": "0.0.20",
-    "@waku/utils": "0.0.14",
-    "debug": "^4.3.4",
-    "dns-query": "^0.11.2",
-    "hi-base32": "^0.5.1",
-    "uint8arrays": "^5.0.1"
-  },
-  "devDependencies": {
-    "@libp2p/peer-id": "^4.0.4",
-    "@libp2p/peer-id-factory": "^4.0.5",
-    "@multiformats/multiaddr": "^12.0.0",
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-json": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@types/chai": "^4.3.11",
-    "@waku/build-utils": "*",
-    "@waku/interfaces": "0.0.21",
-    "chai": "^4.3.10",
-    "cspell": "^8.3.2",
-    "mocha": "^10.3.0",
-    "npm-run-all": "^4.1.5",
-    "rollup": "^4.12.0"
-  },
-  "files": [
-    "dist",
-    "bundle",
-    "src/**/*.ts",
-    "!**/*.spec.*",
-    "!**/*.json",
-    "CHANGELOG.md",
-    "LICENSE",
-    "README.md"
-  ]
-}
