--- packages/dns-discovery/tsconfig.dev.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../tsconfig.dev"
-}
