--- packages/dns-discovery/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig",
-  "compilerOptions": {
-    "outDir": "dist/",
-    "rootDir": "src",
-    "tsBuildInfoFile": "dist/.tsbuildinfo"
-  },
-  "include": ["src"],
-  "exclude": ["src/**/*.spec.ts", "src/test_utils"]
-}
