--- packages/dns-discovery/tsconfig.karma.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "module": "esnext"
-  },
-  "exclude": []
-}
