+++ packages/sdk/package.json
@@ -67,7 +67,7 @@
     "@libp2p/ping": "^1.0.11",
     "@libp2p/websockets": "^8.0.11",
     "@waku/core": "0.0.26",
-    "@waku/dns-discovery": "0.0.20",
+    "@waku/discovery": "0.0.1",
     "@waku/interfaces": "0.0.21",
     "@waku/local-peer-cache-discovery": "^0.0.1",
     "@waku/peer-exchange": "^0.0.19",
